Reorg of utils into separate modules
